Serde(UT): separate serializer tests with macros.
